Move column mapping to code client
