+++ src/promptflow-evals/tests/evals/unittests/test_evaluate.py
@@ -11,10 +11,10 @@ from promptflow.client import PFClient
 from promptflow.evals._constants import DEFAULT_EVALUATION_RESULTS_FILE_NAME
 from promptflow.evals.evaluate import evaluate
 from promptflow.evals.evaluate._evaluate import (
-    _apply_column_mapping,
     _apply_target_to_data,
     _rename_columns_conditionally
     )
+from promptflow.evals.evaluate._utils import _apply_column_mapping
 from promptflow.evals.evaluators import F1ScoreEvaluator, GroundednessEvaluator
